requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.1
blis==0.4.1
boto==2.49.0
boto3==1.10.44
botocore==1.13.44
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cymem==2.0.3
decorator==4.4.1
docutils==0.15.2
en-core-web-sm==2.2.0
gensim==3.8.1
get==2019.4.13
glove-python==0.1.0
idna==2.8
importlib-metadata==1.3.0
infomap==1.0.0b54
isort==4.3.21
jmespath==0.9.4
joblib==0.14.1
langdetect==1.0.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
murmurhash==1.0.2
networkx==2.4
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
plac==1.1.3
post==2019.4.13
preshed==3.0.2
public==2019.4.13
pyemd==0.5.1
pylint==2.4.4
pymongo==3.9.0
pyspellchecker==0.5.3
python-dateutil==2.8.1
python-louvain==0.13
PyYAML==5.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
stop-words==2018.7.23
thinc==7.3.1
tqdm==4.40.2
TwitterAPI==2.5.10
typed-ast==1.4.0
urllib3==1.25.7
wasabi==0.4.2
wrapt==1.11.2
zipp==0.6.0
